refactor(parser): move handlers into separate file
